+++ sdk/keyvault/Azure.Security.KeyVault.Keys/tests/KeyClientTests.cs
@@ -265,10 +265,10 @@ namespace Azure.Security.KeyVault.Keys.Tests
         public void GetKeyRotationPolicyValidation()
         {
             ArgumentException ex = Assert.ThrowsAsync<ArgumentNullException>(async () => await Client.GetKeyRotationPolicyAsync(null));
-            Assert.AreEqual("name", ex.ParamName);
+            Assert.AreEqual("keyName", ex.ParamName);
 
             ex = Assert.ThrowsAsync<ArgumentException>(async () => await Client.GetKeyRotationPolicyAsync(string.Empty));
-            Assert.AreEqual("name", ex.ParamName);
+            Assert.AreEqual("keyName", ex.ParamName);
         }
 
         [Test]
@@ -285,10 +285,10 @@ namespace Azure.Security.KeyVault.Keys.Tests
         public void UpdateKeyRotationPolicyValidation()
         {
             ArgumentException ex = Assert.ThrowsAsync<ArgumentNullException>(async () => await Client.UpdateKeyRotationPolicyAsync(null, null));
-            Assert.AreEqual("name", ex.ParamName);
+            Assert.AreEqual("keyName", ex.ParamName);
 
             ex = Assert.ThrowsAsync<ArgumentException>(async () => await Client.UpdateKeyRotationPolicyAsync(string.Empty, null));
-            Assert.AreEqual("name", ex.ParamName);
+            Assert.AreEqual("keyName", ex.ParamName);
 
             ex = Assert.ThrowsAsync<ArgumentNullException>(async () => await Client.UpdateKeyRotationPolicyAsync("test", null));
             Assert.AreEqual("policy", ex.ParamName);